tools/python/requirements.txt
google-genai
google-auth
google-auth-oauthlib
google-api-python-client
requests
